requirements.txt
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
